+++ tests/unit/pywbemcli/test_pywbemcli_operations.py
@@ -30,8 +30,8 @@ import pytest
 import pywbem
 
 from pywbemtools.pywbemcli._pywbemcli_operations import PYWBEMCLIFakedConnection
-from pywbemtools._utils import ensure_unicode, \
-    DEFAULT_CONNECTIONS_FILE
+from pywbemtools._utils import ensure_unicode, DEFAULT_CONNECTIONS_DIR, \
+    DEFAULT_CONNECTIONS_FILE, MOCKCACHE_ROOT_DIR
 from pywbemtools.pywbemcli.mockscripts import DeprecatedSetupWarning, \
     SetupNotSupportedError
 
@@ -111,12 +111,10 @@ def get_mockcache_dir(connection_name):
 
     We assume the connection name is unique across all connection files.
     """
-    mockcache_rootdir = os.path.join(os.path.expanduser('~'),
-                                     '.pywbemcli_mockcache')
-    if os.path.isdir(mockcache_rootdir):
-        for _dir in os.listdir(mockcache_rootdir):
+    if os.path.isdir(DEFAULT_CONNECTIONS_DIR):
+        for _dir in os.listdir(MOCKCACHE_ROOT_DIR):
             if _dir.endswith('.' + connection_name):
-                dir_path = os.path.join(mockcache_rootdir, _dir)
+                dir_path = os.path.join(MOCKCACHE_ROOT_DIR, _dir)
                 return dir_path
     raise ValueError